Rename private _ files to public
