+++ packages/react-impulse-form/src/impulse-form-switch/_impulse-form-switch-params.ts
@@ -1,20 +1,20 @@
 import type { ImpulseForm } from "../impulse-form/impulse-form"
 import type { ImpulseFormParams } from "../impulse-form/impulse-form-params"
 
-import type { ImpulseFormSwitchError } from "./_impulse-form-switch-error"
-import type { ImpulseFormSwitchErrorSetter } from "./_impulse-form-switch-error-setter"
-import type { ImpulseFormSwitchErrorVerbose } from "./_impulse-form-switch-error-verbose"
-import type { ImpulseFormSwitchFlag } from "./_impulse-form-switch-flag"
-import type { ImpulseFormSwitchFlagSetter } from "./_impulse-form-switch-flag-setter"
-import type { ImpulseFormSwitchFlagVerbose } from "./_impulse-form-switch-flag-verbose"
-import type { ImpulseFormSwitchInput } from "./_impulse-form-switch-input"
-import type { ImpulseFormSwitchInputSetter } from "./_impulse-form-switch-input-setter"
-import type { ImpulseFormSwitchOutput } from "./_impulse-form-switch-output"
-import type { ImpulseFormSwitchOutputVerbose } from "./_impulse-form-switch-output-verbose"
-import type { ImpulseFormSwitchValidateOn } from "./_impulse-form-switch-validate-on"
-import type { ImpulseFormSwitchValidateOnSetter } from "./_impulse-form-switch-validate-on-setter"
-import type { ImpulseFormSwitchValidateOnVerbose } from "./_impulse-form-switch-validate-on-verbose"
 import type { ImpulseFormSwitchBranches } from "./impulse-form-switch-branches"
+import type { ImpulseFormSwitchError } from "./impulse-form-switch-error"
+import type { ImpulseFormSwitchErrorSetter } from "./impulse-form-switch-error-setter"
+import type { ImpulseFormSwitchErrorVerbose } from "./impulse-form-switch-error-verbose"
+import type { ImpulseFormSwitchFlag } from "./impulse-form-switch-flag"
+import type { ImpulseFormSwitchFlagSetter } from "./impulse-form-switch-flag-setter"
+import type { ImpulseFormSwitchFlagVerbose } from "./impulse-form-switch-flag-verbose"
+import type { ImpulseFormSwitchInput } from "./impulse-form-switch-input"
+import type { ImpulseFormSwitchInputSetter } from "./impulse-form-switch-input-setter"
+import type { ImpulseFormSwitchOutput } from "./impulse-form-switch-output"
+import type { ImpulseFormSwitchOutputVerbose } from "./impulse-form-switch-output-verbose"
+import type { ImpulseFormSwitchValidateOn } from "./impulse-form-switch-validate-on"
+import type { ImpulseFormSwitchValidateOnSetter } from "./impulse-form-switch-validate-on-setter"
+import type { ImpulseFormSwitchValidateOnVerbose } from "./impulse-form-switch-validate-on-verbose"
 
 export interface ImpulseFormSwitchParams<
   TKind extends ImpulseForm,
