+++ packages/react-impulse-form/src/impulse-form-switch/_impulse-form-switch-state.ts
@@ -20,23 +20,23 @@ import type { ValidateStrategy } from "../validate-strategy"
 
 import { ImpulseFormSwitch } from "./_impulse-form-switch"
 import type { ImpulseFormSwitchConciseParam } from "./_impulse-form-switch-concise-param"
-import type { ImpulseFormSwitchError } from "./_impulse-form-switch-error"
-import type { ImpulseFormSwitchErrorSetter } from "./_impulse-form-switch-error-setter"
-import type { ImpulseFormSwitchErrorVerbose } from "./_impulse-form-switch-error-verbose"
-import type { ImpulseFormSwitchFlag } from "./_impulse-form-switch-flag"
-import type { ImpulseFormSwitchFlagSetter } from "./_impulse-form-switch-flag-setter"
-import type { ImpulseFormSwitchFlagVerbose } from "./_impulse-form-switch-flag-verbose"
-import type { ImpulseFormSwitchInput } from "./_impulse-form-switch-input"
-import type { ImpulseFormSwitchInputSetter } from "./_impulse-form-switch-input-setter"
-import type { ImpulseFormSwitchOutput } from "./_impulse-form-switch-output"
-import type { ImpulseFormSwitchOutputVerbose } from "./_impulse-form-switch-output-verbose"
 import type { ImpulseFormSwitchParams } from "./_impulse-form-switch-params"
-import type { ImpulseFormSwitchValidateOn } from "./_impulse-form-switch-validate-on"
-import type { ImpulseFormSwitchValidateOnSetter } from "./_impulse-form-switch-validate-on-setter"
-import type { ImpulseFormSwitchValidateOnVerbose } from "./_impulse-form-switch-validate-on-verbose"
-import type { ImpulseFormSwitchVerboseParam } from "./_impulse-form-switch-verbose-param"
 import type { ImpulseFormSwitchBranch } from "./impulse-form-switch-branch"
 import type { ImpulseFormSwitchBranches } from "./impulse-form-switch-branches"
+import type { ImpulseFormSwitchError } from "./impulse-form-switch-error"
+import type { ImpulseFormSwitchErrorSetter } from "./impulse-form-switch-error-setter"
+import type { ImpulseFormSwitchErrorVerbose } from "./impulse-form-switch-error-verbose"
+import type { ImpulseFormSwitchFlag } from "./impulse-form-switch-flag"
+import type { ImpulseFormSwitchFlagSetter } from "./impulse-form-switch-flag-setter"
+import type { ImpulseFormSwitchFlagVerbose } from "./impulse-form-switch-flag-verbose"
+import type { ImpulseFormSwitchInput } from "./impulse-form-switch-input"
+import type { ImpulseFormSwitchInputSetter } from "./impulse-form-switch-input-setter"
+import type { ImpulseFormSwitchOutput } from "./impulse-form-switch-output"
+import type { ImpulseFormSwitchOutputVerbose } from "./impulse-form-switch-output-verbose"
+import type { ImpulseFormSwitchValidateOn } from "./impulse-form-switch-validate-on"
+import type { ImpulseFormSwitchValidateOnSetter } from "./impulse-form-switch-validate-on-setter"
+import type { ImpulseFormSwitchValidateOnVerbose } from "./impulse-form-switch-validate-on-verbose"
+import type { ImpulseFormSwitchVerboseParam } from "./impulse-form-switch-verbose-param"
 
 type ImpulseFormSwitchStateBranches<TBranches> = {
   [TBranch in keyof TBranches]: ImpulseFormState<
