+++ packages/react-impulse-form/src/impulse-form-switch/impulse-form-switch-error-verbose.ts
@@ -1,9 +1,9 @@
 import type { ImpulseForm } from "../impulse-form/impulse-form"
 
-import type { ImpulseFormSwitchVerboseParam } from "./_impulse-form-switch-verbose-param"
 import type { ImpulseFormSwitchBranches } from "./impulse-form-switch-branches"
+import type { ImpulseFormSwitchVerboseParam } from "./impulse-form-switch-verbose-param"
 
-export type ImpulseFormSwitchValidateOnVerbose<
+export type ImpulseFormSwitchErrorVerbose<
   TKind extends ImpulseForm,
   TBranches extends ImpulseFormSwitchBranches<TKind>,
-> = ImpulseFormSwitchVerboseParam<TKind, TBranches, "validateOn.schema.verbose">
+> = ImpulseFormSwitchVerboseParam<TKind, TBranches, "error.schema.verbose">
