+++ packages/react-impulse-form/src/impulse-form-switch/impulse-form-switch-flag-setter.ts
@@ -2,34 +2,33 @@ import type { Setter } from "~/tools/setter"
 
 import type { GetImpulseFormParam } from "../impulse-form/get-impulse-form-param"
 import type { ImpulseForm } from "../impulse-form/impulse-form"
-import type { ValidateStrategy } from "../validate-strategy"
 
 import type { GetImpulseFormSwitchBranchesParam } from "./_get-impulse-form-switch-branches-param"
 import type { ImpulseFormSwitchBranchUnion } from "./_impulse-form-switch-branch-union"
 import type { ImpulseFormSwitchConciseSchema } from "./_impulse-form-switch-concise-schema"
-import type { ImpulseFormSwitchValidateOnVerbose } from "./_impulse-form-switch-validate-on-verbose"
-import type { ImpulseFormSwitchVerboseSchema } from "./_impulse-form-switch-verbose-schema"
 import type { ImpulseFormSwitchBranches } from "./impulse-form-switch-branches"
+import type { ImpulseFormSwitchFlagVerbose } from "./impulse-form-switch-flag-verbose"
+import type { ImpulseFormSwitchVerboseSchema } from "./impulse-form-switch-verbose-schema"
 
-export type ImpulseFormSwitchValidateOnSetter<
+export type ImpulseFormSwitchFlagSetter<
   TKind extends ImpulseForm,
   TBranches extends ImpulseFormSwitchBranches<TKind>,
 > = Setter<
   // concise
-  | ValidateStrategy
+  | boolean
 
   // concise details
   | Partial<
       ImpulseFormSwitchConciseSchema<
-        GetImpulseFormParam<TKind, "validateOn.setter">,
+        GetImpulseFormParam<TKind, "flag.setter">,
         Setter<
-          | ValidateStrategy
-          | ImpulseFormSwitchBranchUnion<TKind, TBranches, "validateOn.setter">,
+          | boolean
+          | ImpulseFormSwitchBranchUnion<TKind, TBranches, "flag.setter">,
           [
             ImpulseFormSwitchBranchUnion<
               TKind,
               TBranches,
-              "validateOn.schema.verbose"
+              "flag.schema.verbose"
             >,
           ]
         >
@@ -39,23 +38,18 @@ export type ImpulseFormSwitchValidateOnSetter<
   // verbose
   | Partial<
       ImpulseFormSwitchVerboseSchema<
-        GetImpulseFormParam<TKind, "validateOn.setter">,
+        GetImpulseFormParam<TKind, "flag.setter">,
         Setter<
-          | ValidateStrategy
+          | boolean
           | Partial<
-              GetImpulseFormSwitchBranchesParam<TBranches, "validateOn.setter">
+              GetImpulseFormSwitchBranchesParam<TBranches, "flag.setter">
             >,
-          [
-            GetImpulseFormSwitchBranchesParam<
-              TBranches,
-              "validateOn.schema.verbose"
-            >,
-          ]
+          [GetImpulseFormSwitchBranchesParam<TBranches, "flag.schema.verbose">]
         >
       >
     >,
   [
     // the only argument is the verbose schema
-    ImpulseFormSwitchValidateOnVerbose<TKind, TBranches>,
+    ImpulseFormSwitchFlagVerbose<TKind, TBranches>,
   ]
 >
