+++ packages/react-impulse-form/src/impulse-form-switch/impulse-form-switch-input-setter.ts
@@ -4,9 +4,9 @@ import type { GetImpulseFormParam } from "../impulse-form/get-impulse-form-param
 import type { ImpulseForm } from "../impulse-form/impulse-form"
 
 import type { GetImpulseFormSwitchBranchesParam } from "./_get-impulse-form-switch-branches-param"
-import type { ImpulseFormSwitchInput } from "./_impulse-form-switch-input"
-import type { ImpulseFormSwitchVerboseSchema } from "./_impulse-form-switch-verbose-schema"
 import type { ImpulseFormSwitchBranches } from "./impulse-form-switch-branches"
+import type { ImpulseFormSwitchInput } from "./impulse-form-switch-input"
+import type { ImpulseFormSwitchVerboseSchema } from "./impulse-form-switch-verbose-schema"
 
 export type ImpulseFormSwitchInputSetter<
   TKind extends ImpulseForm,
