+++ packages/react-impulse-form/src/impulse-form-switch/impulse-form-switch-output-verbose.ts
@@ -1,7 +1,7 @@
 import type { ImpulseForm } from "../impulse-form/impulse-form"
 
-import type { ImpulseFormSwitchVerboseParam } from "./_impulse-form-switch-verbose-param"
 import type { ImpulseFormSwitchBranches } from "./impulse-form-switch-branches"
+import type { ImpulseFormSwitchVerboseParam } from "./impulse-form-switch-verbose-param"
 
 export type ImpulseFormSwitchOutputVerbose<
   TKind extends ImpulseForm,
