+++ packages/react-impulse-form/src/impulse-form-switch/impulse-form-switch-output.ts
