+++ packages/react-impulse-form/src/impulse-form-switch/impulse-form-switch-verbose-schema.ts
