+++ packages/react-impulse-form/src/impulse-form-switch/index.ts
@@ -3,24 +3,20 @@ export * from "./is-impulse-form-switch"
 export type * from "./impulse-form-switch-branch"
 export type * from "./impulse-form-switch-branches"
 
-export type * from "./_impulse-form-switch-error-setter"
-export type * from "./_impulse-form-switch-error-verbose"
-export type * from "./_impulse-form-switch-error"
+export type * from "./impulse-form-switch-error-setter"
+export type * from "./impulse-form-switch-error-verbose"
+export type * from "./impulse-form-switch-error"
 
-export type * from "./_impulse-form-switch-flag-setter"
-export type * from "./_impulse-form-switch-flag-verbose"
-export type * from "./_impulse-form-switch-flag"
+export type * from "./impulse-form-switch-flag-setter"
+export type * from "./impulse-form-switch-flag-verbose"
+export type * from "./impulse-form-switch-flag"
 
-export type * from "./_impulse-form-switch-input-setter"
-export type { ImpulseFormSwitchInput } from "./_impulse-form-switch-input"
+export type * from "./impulse-form-switch-input-setter"
+export type * from "./impulse-form-switch-input"
 
-export type * from "./_impulse-form-switch-output-verbose"
-export type * from "./_impulse-form-switch-output"
+export type * from "./impulse-form-switch-output-verbose"
+export type * from "./impulse-form-switch-output"
 
-export type * from "./_impulse-form-switch-validate-on-setter"
-export type * from "./_impulse-form-switch-validate-on-verbose"
-export type * from "./_impulse-form-switch-validate-on"
-
-console.log(
-  "TODO export publicly accesable types as Input, Output, setters, etc",
-)
+export type * from "./impulse-form-switch-validate-on-setter"
+export type * from "./impulse-form-switch-validate-on-verbose"
+export type * from "./impulse-form-switch-validate-on"
